pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spot-the-scam"
version = "0.1.0"
description = "Uncertainty-aware classifier for fraudulent job postings."
authors = [
  {name = "Spot the Scam Team"}
]
requires-python = ">=3.9"
dependencies = [
  "pandas>=2.0.0",
  "numpy>=1.23.0",
  "scipy>=1.10.0",
  "joblib>=1.3.0",
  "beautifulsoup4>=4.12.0",
  "scikit-learn>=1.2.0",
  "lightgbm>=4.0.0",
  "xgboost>=1.7.0",
  "matplotlib>=3.7.0",
  "seaborn>=0.12.0",
  "plotly>=5.15.0",
  "shap>=0.43.0",
  "lime>=0.2.0",
  "pyyaml>=6.0",
  "tqdm>=4.65.0",
  "typer>=0.9.0",
  "rich>=13.0.0",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "tabulate>=0.9.0",
  "pydantic>=2.6.0",
  "transformers>=4.38.0",
  "datasets>=2.16.0",
  "accelerate>=0.25.0",
  "evaluate>=0.4.1",
  "sentencepiece>=0.1.99",
  "torch>=2.2.0",
  "torchvision>=0.17.0",
  "torchaudio>=2.2.0",
  "safetensors>=0.4.2",
  "imbalanced-learn>=0.12.0",
  "kaggle>=1.6.0",
  "diskcache>=5.6.1",
  "jinja2>=3.1.2",
  "orjson>=3.9.15",
  "httpx>=0.27.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
  "pre-commit>=3.6.0",
  "black>=24.1.0",
  "ruff>=0.3.0",
  "mypy>=1.8.0"
]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "H", "B", "UP"]
ignore = ["E203"]

[tool.coverage.run]
branch = true
source = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=spot_scam --cov-report=term-missing"
